Refactored code slightly to be able to access models dir when importing as library
